requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
autoevals==0.1.0
braintrust==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.4
chevron==0.14.0
distro==1.9.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
openai==2.24.0
orjson==3.11.7
packaging==24.2
pinecone==8.1.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
polyleven==0.11.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sseclient-py==1.9.0
text-unidecode==1.3
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.1
